+++ .changeset/plenty-clouds-divide.md
@@ -1,0 +1,7 @@
+---
+'@keystone-next/cloudinary': major
+'@keystone-next/fields-document': major
+'@keystone-next/keystone': major
+---
+
+Moved `graphql` export of `@keystone-next/keystone/types` to `@keystone-next/keystone`
