+++ examples/custom-field/stars-field/index.ts
@@ -5,9 +5,9 @@ import {
   FieldTypeFunc,
   CommonFieldConfig,
   orderDirectionEnum,
-  graphql,
   filters,
 } from '@keystone-next/keystone/types';
+import { graphql } from '@keystone-next/keystone';
 
 // this field is based on the integer field
 // but with validation to ensure the value is within an expected range
