+++ packages/keystone/src/fields/types/checkbox/index.ts
@@ -5,9 +5,9 @@ import {
   fieldType,
   FieldTypeFunc,
   orderDirectionEnum,
-  graphql,
   filters,
 } from '../../../types';
+import { graphql } from '../../..';
 import { assertCreateIsNonNullAllowed, assertReadIsNonNullAllowed } from '../../non-null-graphql';
 import { resolveView } from '../../resolve-view';
