+++ packages/keystone/src/fields/types/float/index.ts
@@ -4,11 +4,11 @@ import {
   BaseGeneratedListTypes,
   FieldTypeFunc,
   CommonFieldConfig,
-  graphql,
   fieldType,
   orderDirectionEnum,
   filters,
 } from '../../../types';
+import { graphql } from '../../..';
 import { assertCreateIsNonNullAllowed, assertReadIsNonNullAllowed } from '../../non-null-graphql';
 import { resolveView } from '../../resolve-view';
