+++ packages/keystone/src/fields/types/text/index.ts
@@ -3,11 +3,11 @@ import {
   BaseGeneratedListTypes,
   CommonFieldConfig,
   fieldType,
-  graphql,
   orderDirectionEnum,
   FieldTypeFunc,
   filters,
 } from '../../../types';
+import { graphql } from '../../..';
 import { assertCreateIsNonNullAllowed, assertReadIsNonNullAllowed } from '../../non-null-graphql';
 import { resolveView } from '../../resolve-view';
