+++ packages/keystone/src/index.ts
@@ -1,2 +1,3 @@
 export { list, gql, graphQLSchemaExtension, config } from './schema/schema';
 export type { ListSchemaConfig, ListConfig, ExtendGraphqlSchema, BaseFields } from './types';
+export { graphql } from './types/schema';
