+++ packages/keystone/src/types/json-field-type-polyfill-for-sqlite.ts
@@ -1,8 +1,8 @@
+import { graphql } from './schema';
 import {
   JSONValue,
   ItemRootValue,
   KeystoneContext,
-  graphql,
   UpdateFieldInputArg,
   ScalarDBField,
   CreateFieldInputArg,
