+++ packages/keystone/src/types/next-fields.ts
@@ -1,8 +1,8 @@
 import Decimal from 'decimal.js';
+import { graphql } from '..';
 import { BaseGeneratedListTypes } from './utils';
 import { CommonFieldConfig } from './config';
 import { DatabaseProvider, FieldDefaultValue } from './core';
-import { graphql } from './schema';
 import { AdminMetaRootVal, JSONValue, KeystoneContext, MaybePromise } from '.';
 
 export { Decimal };
